// blood/blood-frontend/src/shared/api/generated/model/createEmergencyRequestBloodGroup.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Blood Donation Management API
 * Enterprise-ready modular monolith API for blood donation workflows, geo search, and SEO metadata.
 * OpenAPI spec version: 1.0.0
 */

export type CreateEmergencyRequestBloodGroup = typeof CreateEmergencyRequestBloodGroup[keyof typeof CreateEmergencyRequestBloodGroup];


export const CreateEmergencyRequestBloodGroup = {
  A_POSITIVE: 'A_POSITIVE',
  A_NEGATIVE: 'A_NEGATIVE',
  B_POSITIVE: 'B_POSITIVE',
  B_NEGATIVE: 'B_NEGATIVE',
  AB_POSITIVE: 'AB_POSITIVE',
  AB_NEGATIVE: 'AB_NEGATIVE',
  O_POSITIVE: 'O_POSITIVE',
  O_NEGATIVE: 'O_NEGATIVE',
} as const;