// blood/blood-frontend/src/shared/api/generated/model/createEmergencyRequestComponentType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Blood Donation Management API
 * Enterprise-ready modular monolith API for blood donation workflows, geo search, and SEO metadata.
 * OpenAPI spec version: 1.0.0
 */

export type CreateEmergencyRequestComponentType = typeof CreateEmergencyRequestComponentType[keyof typeof CreateEmergencyRequestComponentType];


export const CreateEmergencyRequestComponentType = {
  WHOLE_BLOOD: 'WHOLE_BLOOD',
  RBC: 'RBC',
  PLASMA: 'PLASMA',
  PLATELET: 'PLATELET',
  FFP: 'FFP',
  CRYOPRECIPITATE: 'CRYOPRECIPITATE',
} as const;